# src/docs/swagger.yaml
basePath: /v1
definitions:
...
  main.LoginWithEmailRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  main.LogoutRequest:
    properties:
      access_token:
        type: string
    required:
    - access_token
    type: object
  main.LogoutResponse:
    properties:
      success:
        type: boolean
    type: object
...
  main.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  main.RefreshTokenResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      id_token:
        type: string
      refresh_token:
        type: string
    type: object
  main.RequestRecoveryPassword:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  main.RequestRecoveryPasswordResponse:
    properties:
      success:
        type: boolean
    type: object
...
  main.ResetPasswordRequest:
    properties:
      code:
        type: string
      email:
        type: string
      password:
        type: string
    required:
    - code
    - email
    - password
    type: object
  main.ResetPasswordResponse:
    properties:
      success:
        type: boolean
    type: object
...
  main.TokenResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      id_token:
        type: string
      refresh_token:
        type: string
    type: object
...
  utils.Response-any:
    properties:
      data: {}
      exception:
        type: string
        x-nullable: "true"
      message:
        type: string
      status_code:
        type: integer
    type: object
...
  utils.Response-main_LogoutResponse:
    properties:
      data:
        $ref: '#/definitions/main.LogoutResponse'
      exception:
        type: string
        x-nullable: "true"
      message:
        type: string
      status_code:
        type: integer
    type: object
  utils.Response-main_RefreshTokenResponse:
    properties:
      data:
        $ref: '#/definitions/main.RefreshTokenResponse'
      exception:
        type: string
        x-nullable: "true"
      message:
        type: string
      status_code:
        type: integer
    type: object
  utils.Response-main_RequestRecoveryPasswordResponse:
    properties:
      data:
        $ref: '#/definitions/main.RequestRecoveryPasswordResponse'
      exception:
        type: string
        x-nullable: "true"
      message:
        type: string
      status_code:
        type: integer
    type: object
...
  utils.Response-main_ResetPasswordResponse:
    properties:
      data:
        $ref: '#/definitions/main.ResetPasswordResponse'
      exception:
        type: string
        x-nullable: "true"
      message:
        type: string
      status_code:
        type: integer
    type: object
...
  utils.Response-main_TokenResponse:
    properties:
      data:
        $ref: '#/definitions/main.TokenResponse'
      exception:
        type: string
        x-nullable: "true"
      message:
        type: string
      status_code:
        type: integer
    type: object
...
host: api.smartgo.com
info:
  contact: {}
  description: SmartGo API
  title: SmartGo API
  version: "1.0"
paths:
...
  /auth/recovery-password:
    patch:
      consumes:
      - application/json
      description: Reset password. If the code, password and email are valid, the
        password will be reset.
      parameters:
      - description: Reset password request
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response-main_ResetPasswordResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response-any'
      summary: Reset password
      tags:
      - security
    post:
      consumes:
      - application/json
      description: |-
        Request password recovery. If the email is found, the user will receive an email with a code to reset the password.
        In otherwise, the user will receive an error.
      parameters:
      - description: Request password recovery
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.RequestRecoveryPassword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response-main_RequestRecoveryPasswordResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response-any'
      summary: Request password recovery
      tags:
      - security
  /auth/sessions:
    delete:
      consumes:
      - application/json
      description: |-
        This endpoint is used to logout a user. It invalidates the refresh token not the access token.
        The access token could be used until its expiration but you can´t create a new access token with the same refresh token.
      parameters:
      - description: Access token
        in: body
        name: access_token
        required: true
        schema:
          $ref: '#/definitions/main.LogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response-main_LogoutResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response-any'
      summary: Logout
      tags:
      - security
    patch:
      consumes:
      - application/json
      description: Refresh access token
      parameters:
      - description: Refresh token
        in: body
        name: refresh_token
        required: true
        schema:
          $ref: '#/definitions/main.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response-main_RefreshTokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response-any'
      summary: Refresh access token
      tags:
      - security
    post:
      consumes:
      - application/json
      description: |-
        Login with email
        This use case is used to login a user with email and password.
        The user will be logged in if the email and password are correct, otherwise an error will be returned.
        The token is a JWT token that is used to authenticate the user.
        The token is valid for 1 hour.
      parameters:
      - description: Request body
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/main.LoginWithEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.Response-main_TokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.Response-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/utils.Response-any'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/utils.Response-any'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.Response-any'
      summary: Login with email and password
      tags:
      - security
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"